% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.mirt.R
\name{extract.mirt}
\alias{extract.mirt}
\title{Extract various elements from estimated model objects}
\usage{
extract.mirt(x, what)
}
\arguments{
\item{x}{mirt model of class 'SingleGroupClass', 'MultipleGroupClass', 'MixedClass' or
'DiscreteGroupClass'}

\item{what}{a string indicating what to extract}
}
\description{
A generic function to extract the internal objects from estimated models.
}
\details{

}
\examples{

\dontrun{
mod <- mirt(Science, 1)

extract.mirt(mod, 'logLik')
extract.mirt(mod, 'F')

#multiple group model
grp <- rep(c('G1', 'G2'), each = nrow(Science)/2)
mod2 <- multipleGroup(Science, 1, grp)

grp1 <- extract.group(mod2, 1) #extract single group model
extract.mirt(mod2, 'parvec')
extract.mirt(grp1, 'parvec')

}
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
\doi{10.18637/jss.v048.i06}
}
\seealso{
\code{\link{extract.group}}, \code{\link{extract.item}}, \code{\link{mod2values}}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{extract}
