% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleGroup-methods.R
\docType{methods}
\name{logLik-method}
\alias{logLik-method}
\alias{logLik,SingleGroupClass-method}
\alias{logLik,MultipleGroupClass-method}
\alias{logLik,MixedClass-method}
\alias{logLik,DiscreteClass-method}
\title{Extract log-likelihood}
\usage{
\S4method{logLik}{SingleGroupClass}(object)
}
\arguments{
\item{object}{an object of class \code{SingleGroupClass},
\code{MultipleGroupClass}, or \code{MixedClass}}
}
\description{
Extract the observed-data log-likelihood.
}
\examples{

\dontrun{
x <- mirt(Science, 1)
logLik(x)

}
}
