\name{logLik-methods}
\docType{methods}
\alias{logLik-methods}
\alias{logLik,polymirtClass-method}
\alias{logLik,confmirtClass-method}
\title{ Methods for Function logLik }
\description{
 Methods for function \code{logLik}
}
\section{Methods}{
\describe{Calculate new objects that contain the Monte Carlo estimated observed log-likelihood values for \code{polymirt} and \code{confmirt} objects.

\item{\code{signature(object = "polymirtClass")}}{}

\item{\code{signature(object = "confmirtClass")}}{}
}}

\keyword{methods}

