% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_moments.R
\name{marginal_moments}
\alias{marginal_moments}
\title{Function to calculate the marginal moments for items and bundles fit via IRT models}
\usage{
marginal_moments(
  mod,
  which.items = NULL,
  group = NULL,
  bundle = TRUE,
  density = NULL,
  Theta_lim = c(-6, 6),
  quadpts = 121,
  ...
)
}
\arguments{
\item{mod}{an object of class \code{'SingleGroupClass'} or \code{'MultipleGroupClass'}}

\item{which.items}{vector indicating which items to use in the computation of the expected values.
Default (\code{NULL}) uses all available items}

\item{group}{optional indicator to return only specific group information for multiple group models.
Default compute moments for each group, returning a \code{list}}

\item{bundle}{logical; given \code{which.items}, should the composite of the response functions be used
as a collective bundle? If \code{TRUE} then \code{\link{expected.test}} will be used, otherwise
\code{\link{expected.item}} will be used}

\item{density}{a density function to use for integration. Default assumes the latent traits are from a
normal (Gaussian) distribution. Function definition must be of the form \code{function(quadrature, mean, sd)}
as the values of the mean/variance are extracted and passed from the supplied model}

\item{Theta_lim}{range of integration grid to use when forming expectations}

\item{quadpts}{number of discrete quadrature to use in the computations}

\item{...}{additional arguments passed to the density function}
}
\description{
Given an estimated model and a prior density function, compute the marginal moments for either
and item or a bundle of items. Function returns the first found moments implied by the model
and select density function (MEAN, VAR, SKEW, and KURT). Currently limited to unidimensional IRT models.
}
\examples{

\donttest{

# single group
dat <- expand.table(deAyala)
mod <- mirt(dat)
TS <- rowSums(dat)

# expected moments of total scores given model
marginal_moments(mod)
c(mean=mean(TS), var=var(TS))  # first two moments of data

# same, however focusing on individual items
marginal_moments(mod, bundle = FALSE)
cbind(mean=colMeans(dat), var=apply(dat, 2, var)) # first two moments of data

############################################
## same as above, however with multiple group model

set.seed(1234)
group <- sample(c('G1', 'G2'), nrow(dat), replace=TRUE)
modMG <- multipleGroup(dat, group=group,
 invariance=c(colnames(dat), 'free_mean', 'free_var'))
coef(modMG, simplify=TRUE)

# expected moments of total scores given model
marginal_moments(modMG)
marginal_moments(modMG, group = 'G1') # specific group only

# same, however focusing on individual items
marginal_moments(modMG, bundle = FALSE)


}
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
\doi{10.18637/jss.v048.i06}
}
\seealso{
\code{\link{expected.item}}, \code{\link{expected.test}}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
