% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirt-package.R
\name{mirt-package}
\alias{mirt-package}
\title{Full information maximum likelihood estimation of IRT models.}
\description{
Full information maximum likelihood estimation of multidimensional IRT models
}
\details{
Analysis of dichotomous and polytomous response data using
unidimensional and multidimensional latent trait models under the Item
Response Theory paradigm. Exploratory and confirmatory models can be
estimated with quadrature (EM) or stochastic (MHRM) methods. Confirmatory
bi-factor and two-tier analyses are available for modeling item testlets.
Multiple group analysis and mixed effects designs also are available for
detecting differential item and test functioning as well as modeling
item and person covariates. Finally, latent class models such as the DINA,
DINO, multidimensional latent class, and several other discrete variable
models are supported.

Users interested in the most recent version of this package can visit
\url{https://github.com/philchalmers/mirt} and follow the instructions
for installing the package from source. Questions regarding the package can
be sent to the mirt-package Google Group, located at
\url{https://groups.google.com/forum/#!forum/mirt-package}. User contributed files,
workshop files, and evaluated help files are also available on the package wiki
(\url{https://github.com/philchalmers/mirt/wiki}).
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
\doi{10.18637/jss.v048.i06}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}
