% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00-classes.R
\docType{class}
\name{MultipleGroupClass-class}
\alias{MultipleGroupClass-class}
\title{Class "MultipleGroupClass"}
\description{
Defines the object returned from \code{\link{multipleGroup}}.
}
\section{Slots}{


\describe{
    \item{\code{Call}:}{function call }
    \item{\code{Data}:}{list of data, sometimes in different forms }
    \item{\code{Options}:}{list of estimation options}
    \item{\code{Fit}:}{a list of fit information }
    \item{\code{Model}:}{a list of model-based information }
    \item{\code{ParObjects}:}{a list of the S4 objects used during estimation}
    \item{\code{OptimInfo}:}{a list of arguments from the optimization process}
    \item{\code{Internals}:}{a list of internal arguments for secondary computations (inspecting this
      object is generally not required)}
    \item{\code{vcov}:}{a matrix represented the asymtotic covariance matrix of the parameter estimates}
    \item{\code{time}:}{a data.frame indicating the breakdown of computation times in seconds}
}
}

\section{Methods}{


\describe{
   \item{coef}{\code{signature(object = "MultipleGroupClass")}}
   \item{print}{\code{signature(x = "MultipleGroupClass")} }
   \item{show}{\code{signature(object = "MultipleGroupClass")} }
   \item{anova}{\code{signature(object = "MultipleGroupClass")} }
}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{classes}

