% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirt-package.R
\docType{data}
\name{pirt_DIF}
\alias{pirt_DIF}
\title{Simulated datasets for PIRT-DIF}
\description{
Three associated datasets for PIRT-DIF, stored as a list (more information to come).
}
\examples{

\donttest{
data(pirt_DIF)

# dataset 1
dat1 <- pirt_DIF$dat1
group <- dat1$group
dat <- dat1[,-1]
itemstats(dat, group=group)

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}
