% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-item_methods.R
\name{print.mirt_df}
\alias{print.mirt_df}
\title{Print generic for customized data.frame console output}
\usage{
\method{print}{mirt_df}(x, digits = 3, ...)
}
\arguments{
\item{x}{object of class \code{'mirt_df'}}

\item{digits}{number of digits to round}

\item{...}{additional arguments passed to \code{print(...)}}
}
\description{
Privides a nicer output for most printed \code{data.frame} objects defined by functions in \code{mirt}.
}
