% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-item_methods.R
\name{print.mirt_list}
\alias{print.mirt_list}
\title{Print generic for customized list console output}
\usage{
\method{print}{mirt_list}(x, digits = 3, ...)
}
\arguments{
\item{x}{object of class \code{'mirt_list'}}

\item{digits}{number of digits to round}

\item{...}{additional arguments passed to \code{print(...)}}
}
\description{
Privides a nicer output for most printed \code{list} objects defined by functions in \code{mirt}.
}

