% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mirt_matrix}
\alias{print.mirt_matrix}
\title{Print generic for customized matrix console output}
\usage{
\method{print}{mirt_matrix}(x, digits = 3, ...)
}
\arguments{
\item{x}{object of class \code{'mirt_matrix'}}

\item{digits}{number of digits to round}

\item{...}{additional arguments passed to \code{print(...)}}
}
\description{
Provides a nicer output for most printed \code{matrix}
objects defined by functions in \code{mirt}.
}
