\name{SAT12}
\alias{SAT12}
\title{Description of SAT12 data}
\description{
Description of data sets.
}

\usage{
data(SAT12)
}

 \arguments{
  \item{SAT12}{data obtained from the TESTFACT manual; 32 pattern scored items for a grade 12 science assessment test (SAT) measuring topics of chemistry, biology, and physics. The scoring key for these data is [1, 4, 5, 2, 3, 1, 2, 1, 3, 1, 2, 4, 2, 1, 5, 3, 4, 4, 1, 4, 3, 3, 4, 1, 3, 5, 1, 3, 1, 5, 4, 5], respectively}
 }

\author{Phil Chalmers \email{rphilip.chalmers@gmail.com}}

\keyword{data} 