% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show-method}
\alias{show,DiscreteClass-method}
\alias{show,MixedClass-method}
\alias{show,MultipleGroupClass-method}
\alias{show,SingleGroupClass-method}
\alias{show-method}
\title{Show model object}
\usage{
\S4method{show}{SingleGroupClass}(object)
}
\arguments{
\item{object}{an object of class \code{SingleGroupClass},
  \code{MultipleGroupClass}, or \code{MixedClass}}
}
\description{
Print model object summaries to the console.
}
\examples{
\dontrun{
x <- mirt(Science, 1)
show(x)
}
}

