% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleGroup-methods.R
\docType{methods}
\name{vcov-method}
\alias{vcov-method}
\alias{vcov,SingleGroupClass-method}
\alias{vcov,MixtureClass-method}
\alias{vcov,MultipleGroupClass-method}
\alias{vcov,MixedClass-method}
\alias{vcov,DiscreteClass-method}
\title{Extract parameter variance covariance matrix}
\usage{
\S4method{vcov}{SingleGroupClass}(object)
}
\arguments{
\item{object}{an object of class \code{SingleGroupClass},
\code{MultipleGroupClass}, or \code{MixedClass}}
}
\description{
Extract parameter variance covariance matrix
}
\examples{

\dontrun{
x <- mirt(Science, 1, SE=TRUE)
vcov(x)

}
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
\doi{10.18637/jss.v048.i06}
}
