\docType{data}
\name{deAyala}
\alias{deAyala}
\title{Description of deAyala data}
\description{
  Mathematics data from de Ayala (2009; pg. 14); 5 item
  dataset in table format.
}
\examples{
\dontrun{
dat <- expand.table(deAyala)
head(dat)
}
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\references{
  de Ayala, R. J. (2009). \emph{The theory and practice of
  item response theory}. Guilford Press.
}
\keyword{data}
\keyword{datasets}

