% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estfun.R
\name{estfun.AllModelClass}
\alias{estfun.AllModelClass}
\title{Extract Empirical Estimating Functions}
\usage{
estfun.AllModelClass(object)
}
\arguments{
\item{object}{a computed model object of class \code{SingleGroupClass} or
\code{MultipleGroupClass}}
}
\value{
An n x k matrix corresponding to n observations and k parameters.
}
\description{
A function for extracting the empirical estimating functions of a fitted
\code{\link{mirt}} or \code{\link{multipleGroup}} model. This is the
derivative of the log-likelihood with respect to the parameter vector,
evaluated at the observed (case-wise) data. In other words, this function
returns the case-wise scores, evaluated at the fitted model parameters.
}
\examples{

\dontrun{
mod <- mirt(expand.table(LSAT7), 1, SE = TRUE, SE.type = "crossprod")
(sc <- estfun.AllModelClass(mod))
colSums(sc)
vc <- mod@vcov
all.equal(crossprod(sc), chol2inv(chol(vc)), check.attributes = FALSE)

}
}
\seealso{
\code{\link{mirt}}, \code{\link{multipleGroup}}
}
\author{
Lennart Schneider \email{lennart.sch@web.de}
}
\keyword{empirical}
\keyword{estimating}
\keyword{functions}
