% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedCalib.R
\name{fixedCalib}
\alias{fixedCalib}
\title{Fixed-item calibration method}
\usage{
fixedCalib(
  data,
  model = 1,
  old_mod,
  PAU = "MWU",
  NEMC = "MEM",
  technical = list(),
  ...
)
}
\arguments{
\item{data}{new data to be used for calibration. Note that to be consistent
with the \code{mod} object, observed responses/NA placeholders must be included
to link the item names used in the original \code{mod} definition
(i.e., \code{extract.mirt(mod, what = 'itemnames')})}

\item{old_mod}{a model of class SingleGroupClass fitted using \code{\link{mirt}}}

\item{PAU}{prior ability update (PAU) approach. Supports none (\code{"NWU"}),
one (\code{"OWU"}), and many (\code{"MWU"})}

\item{technical}{list of technical estimation arguments
(see \code{\link{mirt}} for details)}

\item{...}{additional arguments to pass to \code{\link{mirt}}}

\item{NEM}{number of EM cycles (NEMC) to use for the to-be-estimated parameters.
Supports one (\code{"OEM"}) and many (\code{"MEM"})}
}
\description{
Implements the set of fixed-item calibration methods described by Kim (2006). The initial
calibrated model must be fitted via \code{\link{mirt}}, and is currently limited to
unidimensional models only.
}
\examples{
\dontrun{

# single factor
set.seed(12345)
J <- 50
a <- matrix(abs(rnorm(J,1,.3)), ncol=1)
d <- matrix(rnorm(J,0,.7),ncol=1)
itemtype <- rep('2PL', nrow(a))

# calibration data theta ~ N(0,1)
N <- 3000
dataset1 <- simdata(a, d, N, itemtype)

# more extreme to-be-calibrate data (Kim, 2006, p. 365)
dataset2 <- simdata(a, d, N = 1000, itemtype=itemtype,
                    mu=1, sigma=matrix(1.4^2))

# last 40\% of experimental items not given to calibration group
#     (unobserved; hence removed)
dataset1 <- dataset1[,-c(J:(J*.6))]

# assume first 60\% of items not given to new group
dataset2[,colnames(dataset1)] <- NA

# calibrated model
mod <- mirt(dataset1, model = 1)
coef(mod, simplify=TRUE)

#--------------------------------------

# No Prior Weights Updating and One EM Cycle (NWU-OEM)
NWU_OEM <- fixedCalib(dataset2, model = 1, old_mod = mod, PAU = 'NWU', NEMC = 'OEM')
coef(NWU_OEM, simplify=TRUE)
plot(NWU_OEM, type = 'empiricalhist')

# No Prior Weights Updating and Multiple EM Cycles (NWU-MEM)
NWU_MEM <- fixedCalib(dataset2, model = 1, old_mod = mod, PAU = 'NWU')
coef(NWU_MEM, simplify=TRUE)
plot(NWU_MEM, type = 'empiricalhist')

# One Prior Weights Updating and One EM Cycle (OWU-OEM)
OWU_OEM <- fixedCalib(dataset2, model = 1, old_mod = mod, PAU = 'OWU', NEMC = "OEM")
coef(OWU_OEM, simplify=TRUE)
plot(OWU_OEM, type = 'empiricalhist')

# One Prior Weights Updating and Multiple EM Cycles (OWU-MEM)
OWU_MEM <- fixedCalib(dataset2, model = 1, old_mod = mod, PAU = 'OWU')
coef(OWU_MEM, simplify=TRUE)
plot(OWU_MEM, type = 'empiricalhist')

# Multiple Prior Weights Updating and Multiple EM Cycles (MWU-MEM)
MWU_MEM <- fixedCalib(dataset2, model = 1, old_mod = mod)
coef(MWU_MEM, simplify=TRUE)
plot(MWU_MEM, type = 'empiricalhist')

}
}
\references{
Kim, S. (2006). A comparative study of IRT fixed parameter calibration methods.
\emph{Journal of Educational Measurement, 4}(43), 355–381.
}
