\name{fscores-methods}
\docType{methods}
\alias{fscores-methods}
\alias{fscores,bfactorClass-method}
\alias{fscores,mirtClass-method}
\title{ Methods for Function fscores }
\description{
 Save tabulated or full data factor scores for \code{mirt} or \code{bfactor} using EAP or MAP scoring.
}
\section{Methods}{
  \describe{
    \item{fscores}{\code{signature(object = "bfactorClass")}}
    \item{fscores}{\code{signature(object = "mirtClass")}}
  }
}

\author{Phil Chalmers \email{rphilip.chalmers@gmail.com}}

\keyword{methods}

