\name{fscores-methods}
\docType{methods}
\alias{fscores-methods}
\alias{fscores,bfactorClass-method}
\alias{fscores,mirtClass-method}
\title{ Methods for Function fscores }
\description{
 Methods for function \code{fscores} 
}
\section{Methods}{
\describe{Save tabulated or full data factor scores for \code{mirt} or \code{bfactor} using EAP or MAP scoring.

\item{\code{signature(object = "bfactorClass")}}{}

\item{\code{signature(object = "mirtClass")}}{}
}}

\keyword{methods}

