\docType{methods}
\name{fscores-methods}
\alias{fscores-methods}
\alias{fscores,bfactorClass-method}
\alias{fscores,confmirtClass-method}
\alias{fscores,mirtClass-method}
\alias{fscores,polymirtClass-method}
\title{Methods for Function fscores}
\description{
  Save tabulated or full data factor scores for \code{mirt}
  or \code{bfactor} using EAP or MAP scoring.
}
\section{Methods}{
  \describe{ \item{fscores}{\code{signature(object =
  "bfactorClass")}} \item{fscores}{\code{signature(object =
  "mirtClass")}} \item{fscores}{\code{signature(object =
  "polymirtClass")}} \item{fscores}{\code{signature(object
  = "confmirtClass")}} }
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{methods}

