\name{itemplot-methods}
\docType{methods}
\alias{itemplot-methods}
\alias{itemplot,bfactorClass,numeric-method}
\alias{itemplot,mirtClass,numeric-method}
\alias{itemplot,polymirtClass,numeric-method}
\title{ Methods for Function itemplot}
\description{
Methods for function \code{itemplot}  
}
\section{Methods}{
\describe{Plot individual items for fitted \code{mirt}, \code{bfactor}, or \code{polymirt} models. 

\item{\code{signature(object = "bfactorClass", item = "numeric")}}{}

\item{\code{signature(object = "mirtClass", item = "numeric")}}{}

\item{\code{signature(object = "polymirtClass", item = "numeric")}}{}
}}

\keyword{methods}

