\name{itemplot-methods}
\docType{methods}
\alias{itemplot-methods}
\alias{itemplot,bfactorClass,numeric-method}
\alias{itemplot,mirtClass,numeric-method}
\alias{itemplot,polymirtClass,numeric-method}
\title{ Methods for Function itemplot}
\description{
Plot individual items for fitted \code{mirt}, \code{bfactor}, or \code{polymirt} models.
}
\section{Methods}{
  \describe{
  \item{itemplot}{\code{signature(object = "bfactorClass", item = "numeric")}}
  \item{itemplot}{\code{signature(object = "mirtClass", item = "numeric")}}
  \item{itemplot}{\code{signature(object = "polymirtClass", item = "numeric")}}
  }
}

\author{Phil Chalmers \email{rphilip.chalmers@gmail.com}}

\keyword{methods}

