\name{logLik-methods}
\docType{methods}
\alias{logLik-methods}
\alias{logLik,polymirtClass-method}
\alias{logLik,confmirtClass-method}
\title{ Methods for Function logLik }
\description{
 Calculate new objects that contain the Monte Carlo estimated observed log-likelihood values for \code{polymirt} and \code{confmirt} objects.
}
\section{Methods}{
\describe{
\item{logLik}{\code{signature(object = "polymirtClass")}}
\item{logLik}{\code{signature(object = "confmirtClass")}}
}}

\author{Phil Chalmers \email{rphilip.chalmers@gmail.com}}

\keyword{methods}

