% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirt-package.R
\docType{data}
\name{LSAT7}
\alias{LSAT7}
\title{Description of LSAT7 data}
\description{
Data from Bock & Lieberman (1970); contains 5 dichotomously scored
items obtained from the Law School Admissions Test, section 7.

Data from
}
\examples{

\donttest{
dat <- expand.table(LSAT7)
head(dat)
itemstats(dat)

# fit 2PL model for each item
(mod <- mirt(dat))
coef(mod)

# monotonic splines models (see Winsberg, Thissen, and Wainer, 1984)
mod_monospline <- mirt(dat, itemtype = 'monospline')
anova(mod, mod_monospline)
plot(mod_monospline)

# compare item 1 trace-lines
i1 <- extract.item(mod, 1)
i1mono <- extract.item(mod_monospline, 1)
theta <- matrix(seq(-6, 6, length.out=100))
twoPL <- probtrace(i1, theta)[,2]
monospline <- probtrace(i1mono, theta)[,2]

plot(twoPL ~ theta, type = 'l')
lines(monospline ~ theta, col='red')

}

\donttest{
dat <- expand.table(LSAT7)
head(dat)
itemstats(dat)

(mod <- mirt(dat, 1))
coef(mod)
}
}
\references{
Bock, R. D., & Lieberman, M. (1970). Fitting a response model for \emph{n}
dichotomously scored items. \emph{Psychometrika, 35}(2), 179-197.

Bock, R. D., & Lieberman, M. (1970). Fitting a response model for \emph{n}
dichotomously scored items. \emph{Psychometrika, 35}(2), 179-197.
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}
