\name{LSAT7}
\alias{LSAT7}
\title{Description of LSAT7 data}
\description{
Description of data sets.
}

\usage{
data(LSAT7)

}

 \arguments{
   \item{LSAT7}{data from Bock & Lieberman (1970); contains 5 dichotomously scored items obtained from the Law School Admissions Test. }
 }

\references{
Bock, R. D., & Lieberman, M. (1970). Fitting a response model for \emph{n} dichotomously scored items. \emph{Psychometrika, 35}(2), 179-197.
}

\author{Phil Chalmers \email{rphilip.chalmers@gmail.com}}

\keyword{data} 