% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marginal_rxx.R
\name{marginal_rxx}
\alias{marginal_rxx}
\title{Function to calculate the marginal reliability}
\usage{
marginal_rxx(mod, density = dnorm, theta_lim = c(-6, 6), ...)
}
\arguments{
\item{mod}{an object of class \code{'SingleGroupClass'}}

\item{density}{a density function to use for integration. Default assumes the latent traits are from a
normal (Gaussian) distribution}

\item{theta_lim}{a vector containing the range of integration}

\item{...}{additional arguments passed to the density function}
}
\description{
Given an estimated model and a prior density function, compute the marginal reliability. This is only
available for unidimensional tests.
}
\examples{
\dontrun{

dat <- expand.table(deAyala)
mod <- mirt(dat, 1)

# marginal estimate
marginal_rxx(mod)

# empirical estimate (assuming the same prior)
fscores(mod, returnER = TRUE)

# empirical rxx the alternative way, given theta scores and SEs
fs <- fscores(mod, full.scores.SE=TRUE)
head(fs)
empirical_rxx(fs)

}
}
\seealso{
\code{\link{empirical_rxx}}, \code{\link{extract.group}}, \code{\link{testinfo}}
}
\keyword{reliability}

