% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDIFF.R
\name{MDIFF}
\alias{MDIFF}
\title{Compute multidimensional difficulty index}
\usage{
MDIFF(x)
}
\arguments{
\item{x}{an object of class 'SingleGroupClass'}
}
\description{
Returns a matrix containing the MDIFF values (Reckase, 2009).
}
\examples{
\dontrun{

mod <- mirt(Science, 2)
MDIFF(mod)

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\references{
Reckase, M. D. (2009). Multidimensional Item Response Theory. Springer.
}
\seealso{
\code{\link{extract.group}}, \code{\link{MDISC}}
}
\keyword{discrimination}

