% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MDISC.R
\name{MDISC}
\alias{MDISC}
\title{Compute multidimensional discrimination index}
\usage{
MDISC(x)
}
\arguments{
\item{x}{an object of class 'SingleGroupClass'}
}
\description{
Returns a vector containing the MDSIC values (Reckase, 2009).
}
\examples{
\dontrun{

mod <- mirt(Science, 2)
MDISC(mod)

}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\references{
Reckase, M. D. (2009). Multidimensional Item Response Theory. Springer.
}
\seealso{
\code{\link{extract.group}}
}
\keyword{discrimination}

