\docType{package}
\name{mirt-package}
\alias{mirt-package}
\title{Full information maximum likelihood estimation of IRT models.}
\description{
  Full information maximum likelihood estimation of
  multidimensional IRT models
}
\details{
  Analysis of dichotomous and polytomous response data
  using latent trait models under the Item Response Theory
  paradigm. Exploratory models can be estimated via
  quadrature or stochastic methods, a generalized
  confirmatory bi-factor analysis is included, and
  confirmatory models can be fit with a Metropolis-Hastings
  Robbins-Monro algorithm which can include polynomial or
  product constructed latent traits. Multiple group
  analysis and mixed effects designs may be performed for
  unidimensional or multidimensional item response models
  for detecting differential item functioning and modelling
  item and person covariates.

  Users interested in the most recent version of this
  package can visit
  \url{https://github.com/philchalmers/mirt} and follow the
  instructions for installing the package from source.
  Questions regarding the package can be sent to the
  mirt-package Google Group, located at
  \url{https://groups.google.com/forum/#!forum/mirt-package}.
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

