\docType{package}
\name{mirt-package}
\alias{mirt-package}
\title{Full information maximum likelihood estimation of IRT models.}
\description{
  Full information maximum likelihood estimation of
  multidimensional IRT models
}
\details{
  Analysis of dichotomous and polytomous response data
  using latent trait models under the Item Response Theory
  paradigm. Includes univariate and multivariate one-,
  two-, three-, and four-parameter logistic models, graded
  response models, generalized partial credit models,
  nominal models, multiple choice models, and multivariate
  partially-compensatory models. These can be used in an
  exploratory or confirmatory manner with optional user
  defined linear constraints. Exploratory models can be
  estimated via quadrature or stochastic methods, a
  generalized confirmatory bi-factor analysis is included,
  and confirmatory models can be fit with a
  Metropolis-Hastings Robbins-Monro algorithm which can
  include polynomial or product constructed latent traits.
  Additionally, multiple group analysis may be performed
  for unidimensional or multidimensional item response
  models for detecting differential item functioning.

  Users interested in the most recent version of this
  package can visit
  \code{https://github.com/philchalmers/mirt} and follow
  the instructions for installing the package from source.
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

