\docType{package}
\name{mirt-package}
\alias{mirt-package}
\title{Full information maximum likelihood estimation of IRT models.}
\description{
  Full information maximum likelihood estimation of
  multidimensional IRT models
}
\details{
  Analysis of dichotomous and polychotomous response data
  using latent trait models under the Item Response Theory
  paradigm. Includes the multivariate two- and
  three-parameter logistic models, confirmatory bifactor
  analysis, polytomous confirmatory and exploratory item
  response models, and partially-compensatory item response
  modeling in conjunction with non-linear (product) factor
  relationships.
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{package}

