% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleGroup-methods.R
\docType{methods}
\name{print-method}
\alias{print,DiscreteClass-method}
\alias{print,MixedClass-method}
\alias{print,MultipleGroupClass-method}
\alias{print,SingleGroupClass-method}
\alias{print-method}
\title{Print the model objects}
\usage{
\S4method{print}{SingleGroupClass}(x)
}
\arguments{
\item{x}{an object of class \code{SingleGroupClass},
\code{MultipleGroupClass}, or \code{MixedClass}}
}
\description{
Print model object summaries to the console.
}
\examples{

\dontrun{
x <- mirt(Science, 1)
print(x)
}
}

