% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleGroup-methods.R
\name{print-method}
\alias{print-method}
\alias{print,SingleGroupClass-method}
\alias{print,MultipleGroupClass-method}
\alias{print,MixedClass-method}
\alias{print,DiscreteClass-method}
\alias{print,MixtureClass-method}
\title{Print the model objects}
\usage{
\S4method{print}{SingleGroupClass}(x)
}
\arguments{
\item{x}{an object of class \code{SingleGroupClass},
\code{MultipleGroupClass}, or \code{MixedClass}}
}
\description{
Print model object summaries to the console.
}
\examples{

\donttest{
x <- mirt(Science, 1)
print(x)
}
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
\doi{10.18637/jss.v048.i06}
}
