% Generated by roxygen2 (4.0.1): do not edit by hand
\name{probtrace}
\alias{probtrace}
\title{Function to calculate probability trace lines}
\usage{
probtrace(x, Theta)
}
\arguments{
\item{x}{an extracted internal mirt object containing item information}

\item{Theta}{a vector (unidimensional) or matrix (unidimensional/multidimensional) of latent trait values}
}
\description{
Given an internal mirt object extracted from an estimated model
compute the probability trace lines for all categories.
}
\examples{
\dontrun{
mod <- mirt(Science, 1)
extr.2 <- extract.item(mod, 2)
Theta <- matrix(seq(-4,4, by = .1))
traceline <- probtrace(extr.2, Theta)

head(data.frame(traceline, Theta=Theta))

}
}
\seealso{
\code{\link{extract.item}}
}
\keyword{tracelines}

