\name{Science}
\alias{Science}
\title{Description of Science data}
\description{
Description of data sets.
}

\usage{
data(Science)

}

 \arguments{
   \item{Science}{4 item data set borrowed from \code{ltm} package, first example of the \code{grm} function. See more complete documentation therein. }
 }

\author{Phil Chalmers \email{rphilip.chalmers@gmail.com}}

\keyword{data} 