% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Science}
\alias{Science}
\title{Description of Science data}
\description{
A 4-item data set borrowed from \code{ltm} package in R, first example
of the \code{grm()} function. See more complete documentation therein.
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}

