% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirt-package.R
\docType{data}
\name{Science}
\alias{Science}
\title{Description of Science data}
\description{
A 4-item data set borrowed from \code{ltm} package in R, first example
of the \code{grm()} function. See more complete documentation therein, as
well as Karlheinz and Melich (1992).
}
\examples{

\dontrun{
itemstats(Science)

mod <- mirt(Science, 1)
plot(mod, type = 'trace')
}
}
\references{
Karlheinz, R. and Melich, A. (1992). Euro-Barometer 38.1:
\emph{Consumer Protection and Perceptions of Science and Technology}.
INRA (Europe), Brussels. [computer file]
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}
