\docType{data}
\name{Science}
\alias{Science}
\title{Description of Science data}
\description{
  A 4-item data set borrowed from \code{\link[ltm]{ltm}}
  package, first example of the \code{\link[ltm]{grm}}
  function. See more complete documentation therein.
}
\author{
  Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
\keyword{data}
\keyword{datasets}

