% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleGroup-methods.R
\docType{methods}
\name{show-method}
\alias{show-method}
\alias{show,SingleGroupClass-method}
\alias{show,MultipleGroupClass-method}
\alias{show,MixedClass-method}
\alias{show,DiscreteClass-method}
\title{Show model object}
\usage{
\S4method{show}{SingleGroupClass}(object)
}
\arguments{
\item{object}{an object of class \code{SingleGroupClass},
\code{MultipleGroupClass}, or \code{MixedClass}}
}
\description{
Print model object summaries to the console.
}
\examples{

\dontrun{
x <- mirt(Science, 1)
show(x)
}
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
\doi{10.18637/jss.v048.i06}
}
