% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thetaComb}
\alias{thetaComb}
\title{Create all possible combinations of vector input}
\usage{
thetaComb(theta, nfact)
}
\arguments{
\item{x}{the vector from which all possible combinations should be obtained}

\item{k}{the number of observations (and therefore the number of columns to return in
the matrix of combinations)}
}
\value{
a matrix with all possible combinations
}
\description{
This function constructs all possible k-way combinations of an input vector.
It is primarily useful when used in conjunction with the \code{\link{mdirt}} function,
though users may have other uses for it as well. See \code{\link{expand.grid}} for more
flexible combination formats.
}
\examples{

# all possible joint combinations for the vector -4 to 4
thetaComb(-4:4, 2)

# all possible binary combinations for four observations
thetaComb(c(0,1), 4)

}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory
Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.
\doi{10.18637/jss.v048.i06}
}
\author{
Phil Chalmers \email{rphilip.chalmers@gmail.com}
}
